Move the support for resizing devices
Move the DBus support for the Resize dialog to the device tree
of the automatic partitioning module.
