+++ pyanaconda/modules/storage/partitioning/automatic/automatic_interface.py
@@ -75,53 +75,3 @@ class AutoPartitioningInterface(PartitioningInterface):
         :param passphrase: a string with a passphrase
         """
         self.implementation.set_passphrase(passphrase)
-
-    def RemoveDevice(self, device_name: Str):
-        """Remove a device after removing its dependent devices.
-
-        If the device is protected, do nothing. If the device has
-        protected children, just remove the unprotected ones.
-
-        :param device_name: a name of the device
-        """
-        self.implementation.remove_device(device_name)
-
-    def ShrinkDevice(self, device_name: Str, size: UInt64):
-        """Shrink the size of the device.
-
-        :param device_name: a name of the device
-        :param size: a new size in bytes
-        """
-        self.implementation.shrink_device(device_name, size)
-
-    def IsDevicePartitioned(self, device_name: Str) -> Bool:
-        """Is the specified device partitioned?
-
-        :param device_name: a name of the device
-        :return: True or False
-        """
-        return self.implementation.is_device_partitioned(device_name)
-
-    def GetDevicePartitions(self, device_name: Str) -> List[Str]:
-        """Get partitions of the specified device.
-
-        :param device_name: a name of the device
-        :return: a list of device names
-        """
-        return self.implementation.get_device_partitions(device_name)
-
-    def IsDeviceResizable(self, device_name: Str) -> Bool:
-        """Is the specified device resizable?
-
-        :param device_name: a name of the device
-        :return: True or False
-        """
-        return self.implementation.is_device_resizable(device_name)
-
-    def GetDeviceSizeLimits(self, device_name: Str) -> Tuple[UInt64, UInt64]:
-        """Get size limits of the given device.
-
-        :param device_name: a name of the device
-        :return: a tuple of min and max sizes in bytes
-        """
-        return self.implementation.get_device_size_limits(device_name)
