+++ tests/nosetests/pyanaconda_tests/module_part_automatic_test.py
@@ -20,22 +20,22 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from blivet.devices import StorageDevice, DiskDevice, PartitionDevice
-from blivet.formats import get_format
 from blivet.formats.luks import LUKS2PBKDFArgs
 from blivet.size import Size
 
 from pyanaconda.core.configuration.storage import PartitioningType
 from pyanaconda.modules.common.structures.validation import ValidationReport
+from pyanaconda.modules.storage.partitioning.automatic.resizable_module import \
+    ResizableDeviceTreeModule
 from pyanaconda.storage.partspec import PartSpec
 from tests.nosetests.pyanaconda_tests import patch_dbus_publish_object, check_dbus_property, \
-    check_task_creation
+    check_task_creation, check_dbus_object_creation
 
 from pykickstart.constants import AUTOPART_TYPE_LVM_THINP
 
 from dasbus.typing import *  # pylint: disable=wildcard-import
 from pyanaconda.modules.common.constants.objects import AUTO_PARTITIONING
-from pyanaconda.modules.common.errors.storage import UnavailableStorageError, ProtectedDeviceError
+from pyanaconda.modules.common.errors.storage import UnavailableStorageError
 from pyanaconda.modules.common.structures.partitioning import PartitioningRequest
 from pyanaconda.modules.storage.partitioning.automatic.automatic_module import \
     AutoPartitioningModule
@@ -75,6 +75,13 @@ class AutopartitioningInterfaceTestCase(unittest.TestCase):
     def publication_test(self):
         """Test the DBus representation."""
         self.assertIsInstance(self.module.for_publication(), AutoPartitioningInterface)
+
+    @patch_dbus_publish_object
+    def device_tree_test(self, publisher):
+        """Test the device tree."""
+        self.module.on_storage_changed(Mock())
+        path = self.interface.GetDeviceTree()
+        check_dbus_object_creation(self, path, publisher, ResizableDeviceTreeModule)
 
     def enabled_property_test(self):
         """Test the property enabled."""
@@ -147,145 +154,6 @@ class AutopartitioningInterfaceTestCase(unittest.TestCase):
 
         self.assertNotEqual(self.module.storage, storage)
         self.assertIsNotNone(self.module._storage_playground)
-
-    def remove_device_test(self):
-        """Test RemoveDevice."""
-        self.module.on_storage_changed(create_storage())
-
-        dev1 = StorageDevice(
-            "dev1",
-            exists=False,
-            size=Size("15 GiB"),
-            fmt=get_format("disklabel")
-        )
-        dev2 = StorageDevice(
-            "dev2",
-            exists=False,
-            parents=[dev1],
-            size=Size("6 GiB"),
-            fmt=get_format("ext4")
-        )
-        dev3 = StorageDevice(
-            "dev3",
-            exists=False,
-            parents=[dev1],
-            size=Size("9 GiB"),
-            fmt=get_format("ext4")
-        )
-
-        self.module.storage.devicetree._add_device(dev1)
-        self.module.storage.devicetree._add_device(dev2)
-        self.module.storage.devicetree._add_device(dev3)
-
-        dev1.protected = True
-        with self.assertRaises(ProtectedDeviceError):
-            self.interface.RemoveDevice("dev1")
-
-        self.assertIn(dev1, self.module.storage.devices)
-        self.assertIn(dev2, self.module.storage.devices)
-        self.assertIn(dev3, self.module.storage.devices)
-
-        dev1.protected = False
-        dev2.protected = True
-        self.interface.RemoveDevice("dev1")
-
-        self.assertIn(dev1, self.module.storage.devices)
-        self.assertIn(dev2, self.module.storage.devices)
-        self.assertNotIn(dev3, self.module.storage.devices)
-
-        dev2.protected = False
-        self.interface.RemoveDevice("dev1")
-
-        self.assertNotIn(dev1, self.module.storage.devices)
-        self.assertNotIn(dev2, self.module.storage.devices)
-        self.assertNotIn(dev3, self.module.storage.devices)
-
-    def shrink_device_test(self):
-        """Test ShrinkDevice."""
-        self.module.on_storage_changed(create_storage())
-
-        sda1 = StorageDevice(
-            "sda1",
-            exists=False,
-            size=Size("10 GiB"),
-            fmt=get_format("ext4")
-        )
-        self.module.storage.devicetree._add_device(sda1)
-
-        def resize_device(device, size):
-            device.size = size
-
-        self.module.storage.resize_device = resize_device
-
-        sda1.protected = True
-        with self.assertRaises(ProtectedDeviceError):
-            self.interface.ShrinkDevice("sda1", Size("3 GiB").get_bytes())
-
-        sda1.protected = False
-        self.interface.ShrinkDevice("sda1", Size("3 GiB").get_bytes())
-        self.assertEqual(sda1.size, Size("3 GiB"))
-
-        self.interface.ShrinkDevice("sda1", Size("5 GiB").get_bytes())
-        self.assertEqual(sda1.size, Size("3 GiB"))
-
-    def is_device_partitioned_test(self):
-        """Test IsDevicePartitioned."""
-        self.module.on_storage_changed(create_storage())
-        self._add_device(DiskDevice(
-            "dev1"
-        ))
-        self._add_device(DiskDevice(
-            "dev2",
-            fmt=get_format("disklabel")
-        ))
-
-        self.assertEqual(self.interface.IsDevicePartitioned("dev1"), False)
-        self.assertEqual(self.interface.IsDevicePartitioned("dev2"), True)
-
-    def get_device_partitions_test(self):
-        """Test GetDevicePartitions."""
-        self.module.on_storage_changed(create_storage())
-        dev1 = DiskDevice(
-            "dev1"
-        )
-        self._add_device(dev1)
-
-        dev2 = DiskDevice(
-            "dev2",
-            fmt=get_format("disklabel")
-        )
-        self._add_device(dev2)
-
-        dev3 = PartitionDevice(
-            "dev3"
-        )
-        dev2.add_child(dev3)
-        self._add_device(dev3)
-
-        self.assertEqual(self.interface.GetDevicePartitions("dev1"), [])
-        self.assertEqual(self.interface.GetDevicePartitions("dev2"), ["dev3"])
-        self.assertEqual(self.interface.GetDevicePartitions("dev3"), [])
-
-    def is_device_resizable_test(self):
-        """Test IsDeviceResizable."""
-        self.module.on_storage_changed(create_storage())
-        self._add_device(StorageDevice(
-            "dev1"
-        ))
-        self.assertEqual(self.interface.IsDeviceResizable("dev1"), False)
-
-    def get_device_size_limits_test(self):
-        """Test GetDeviceSizeLimits."""
-        self.module.on_storage_changed(create_storage())
-        self._add_device(StorageDevice(
-            "dev1",
-            fmt=get_format("ext4"),
-            size=Size("10 MiB")
-        ))
-
-        min_size, max_size = self.interface.GetDeviceSizeLimits("dev1")
-        self.assertEqual(min_size, 0)
-        self.assertEqual(max_size, 0)
 
     @patch_dbus_publish_object
     def configure_with_task_test(self, publisher):
